Fix the test `Test_Plus_SingleDigit`
`Test_Plus_SingleDigit` is not supposed to be broken, but it was.
Since we made a bug in the test. Test is supposed to add 2 numbers the sum is still single digit.
But we wrote wrong condition for the if statement, even though the sum is 10 which is 2 digits,
still go through the test logic.

Additionally, the test can check whether the return value has single digit.